+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 
+## 17.09.0-ce (2017-10-13)
+- FEATURE: `rc` command added for **rancher** command line client
+- FEATURE: `tf` command added for **terraform** command line client
+
 ## 17.09.0-ce (2017-10-09)
 - UPDATE: now `version_check` won't not occur in background anymore (avoid parasite message)